+++ src/options.js
@@ -11,9 +11,6 @@ class RecrasOptions {
     }
 
     getAnalyticsEvents() {
-        /*if (!this.getAnalyticsObject()) {
-            throw new Error(this.languageHelper.translate('ERR_NO_ANALYTICS'));
-        }*/
         if (!Array.isArray(this.options.analyticsEvents)) {
             this.options.analyticsEvents = RecrasEventHelper.allEvents();
         }
@@ -31,7 +28,7 @@ class RecrasOptions {
         }
         return this.options.analyticsEvents;
     }
-    getAnalyticsObject() {
+    getAnalytics() {
         return this.options.analytics;
     }
     getApiBase() {