+++ changelog.md
@@ -1,7 +1,7 @@
 # Changelog
 
 ## branch tag-manager
-* Add Google Analytics (Tag Manager) integration
+* Add Google Analytics integration
 
 ## 0.13.2 (2019-01-16)
 * Fix error message related to default attachments when you are logged in to the Recras being used