+++ changelog.md
@@ -1,7 +1,8 @@
 # Changelog
 
-## master
+## 0.13.4 (2019-01-29)
 * Add `recras-input-invalid` class to inputs for products that require an other product
+* Fix minimum booking size
 
 ## 0.13.3 (2019-01-17)
 * Fix required products not working when you are logged in to the Recras being used